% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeSigMatrix.R
\name{AugmentSigMatrix}
\alias{AugmentSigMatrix}
\title{Use a t-test to rank to features for each cell type
 The algorithm will be to add one additional gene from each new cell type
 Record the condition number, and plot those.
 Will only consider adding rows shared by fullData and newData}
\usage{
AugmentSigMatrix(origMatrix, fullData, newData, gList, nGenes = 1:100,
  plotToPDF = TRUE, imputeMissing = TRUE, condTol = 1.01,
  postNorm = FALSE, minSumToRem = NA, addTitle = NULL,
  autoDetectMin = FALSE, calcSpillOver = FALSE)
}
\arguments{
\item{origMatrix}{The original signature matrix}

\item{fullData}{The full data for the signature matrix}

\item{newData}{The new data to add signatures from}

\item{gList}{The ordered list of genes from running rankByT() on newData. NOTE: best genes at the bottom!!}

\item{nGenes}{The number of additional genes to consider (DEFAULT: 1:100)}

\item{plotToPDF}{Plot the output condition numbers to a pdf file. (DEFAULT: TRUE)}

\item{imputeMissing}{Set to TRUE to impute missing values. (DEFAULT: TRUE)}

\item{condTol}{Setting higher tolerances will result in smaller numbers extra genes.  1.00 minimized compliment number, 1.01 should be better.  <1 shouldn't work (DEFAULT: 1.00)}

\item{postNorm}{Set to TRUE to normalize new signatures to match old signatures.  To Do: Redo Kappa curve? (DEFAULT: FALSE)}

\item{minSumToRem}{Set to non-NA to remove any row with the sum(abs(row)) < minSumToRem (DEFAULT: NA)}

\item{addTitle}{An optional string to add to the plot and savefile (DEFAULT: NULL)}

\item{autoDetectMin}{Set to true to automatically detect the first local minima. GOOD PRELIMINARY RESULTS (DEAFULT: FALSE)}

\item{calcSpillOver}{Use the training data to calculate a spillover matrix (DEFAULT: FALSE)}
}
\value{
an augmented cell type signature matrix
}
\description{
newMatData <- AugmentSigMatrix(origMatrix, fullData, newData)
}
