% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onlyDeconAlgorithms.R
\name{buildSpilloverMat}
\alias{buildSpilloverMat}
\title{Build a spillover matrix, i.e. what do purified samples deconvolve as?}
\usage{
buildSpilloverMat(refExpr, geneExpr, method = "DCQ")
}
\arguments{
\item{refExpr}{The deconvolution matrix, e.g. LM22 or MGSM27}

\item{geneExpr}{The full gene expression for purified cell types.  Multiple columns (examples) for each column in the reference expr.}

\item{method}{One of 'DCQ', 'SVMDECON', 'DeconRNASeq' (DEFAULT: DCQ)}
}
\value{
A spillover matrix showing how purified cell types deconvolve
}
\description{
spillExpr <- buildSpilloverMat(refExpr, geneExpr, method='DCQ')
}
