% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onlyDeconAlgorithms.R
\name{clustWspillOver}
\alias{clustWspillOver}
\title{Build clusters based on n-pass spillover matrix}
\usage{
clustWspillOver(sigMatrix, geneExpr)
}
\arguments{
\item{sigMatrix}{The deconvolution matrix, e.g. LM22 or MGSM27}

\item{geneExpr}{The source gene expression matrix used to calculate sigMatrix.}
}
\value{
Cell types grouped by cluster
}
\description{
allClusters <- clustWspillOver(sigMatrix, geneExpr)
}
