% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLoops.R
\name{findConvergenceIter}
\alias{findConvergenceIter}
\title{Find out at which iteration the results converge, i.e. the mean results are stable.}
\usage{
findConvergenceIter(curSeq, changePer = 1, winSize = 5)
}
\arguments{
\item{curSeq}{A sequence of results that generated from each iteration of the loop}

\item{changePer}{The maximum percentage of change allowed}

\item{winSize}{The window size for mean calculation}
}
\value{
The minimum number of iterations needed for the results to converge
}
\description{
Find out at which iteration the results converge, i.e. the mean results are stable.
}
