% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeSigMatrix.R
\name{getLM22cells}
\alias{getLM22cells}
\title{LM22 look up table}
\usage{
getLM22cells()
}
\value{
a map of cell types names
}
\description{
Load a map of cell type names
}
\examples{
cellMap <- getLM22cells()
}
