% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onlyDeconAlgorithms.R
\name{heirarchicalClassify}
\alias{heirarchicalClassify}
\title{Build clusters based on n-pass spillover matrix}
\usage{
heirarchicalClassify(sigMatrix, geneExpr, toPred, heirarchData = NULL)
}
\arguments{
\item{sigMatrix}{The deconvolution matrix, e.g. LM22 or MGSM27}

\item{geneExpr}{The source gene expression matrix used to calculate refExpr}

\item{toPred}{The gene expression to ultimately deconvolve}

\item{heirarchData}{The results of heirarchicalSplit OR heirarchicalSplit.sc (DEFAULT: NULL, ie heirarchicalSplit)}
}
\value{
a mtrix of cell counts
}
\description{
cellCounts <- heirarchicalClassify(sigMatrix, geneExpr, toPred, heirarchData=NULL)
}
\details{
UNDER DEVELOPMENT: USE AT YOUR OWN RISK
}
