% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onlyDeconAlgorithms.R
\name{heirarchicalSplit}
\alias{heirarchicalSplit}
\title{Attempt to deconvolve cell types by building a heriarchy of cell types using
  spillToConvergence to determine cell types that are not signficantly different.
  First deconvolve those clusters of cell types.
  Deconvolution matrices are then built to separate the cell types that formerly could
  not be separated.}
\usage{
heirarchicalSplit(sigMatrix, geneExpr)
}
\arguments{
\item{sigMatrix}{The deconvolution matrix, e.g. LM22 or MGSM27}

\item{geneExpr}{The source gene expression matrix used to calculate sigMatrix}
}
\value{
A list of clusters and a list of signature matrices for breaking those clusters
}
\description{
heirarchData <- heirarchicalSplit(sigMatrix, geneExpr)
}
\details{
UNDER DEVELOPMENT: USE AT YOUR OWN RISK

heirarchData <- heirarchicalSplit(sigMatrix, geneExpr)
}
