\name{Licenses}
\alias{Licenses}
\docType{data}
\title{
Licenses required by Celgene legal
}
\description{
This software is covered by the MIT license.  Celgene legal thought it was wise to break the license up into the two license files included in this list.
}
\usage{data("Licenses")}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\source{
https://www.r-project.org/Licenses/MIT
}
\examples{
data(Licenses)
str(Licenses)
}
\keyword{datasets}
