% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeSigMatrix.R
\name{loadMGSM27}
\alias{loadMGSM27}
\title{Load the MGSM27 signature matrix}
\usage{
loadMGSM27()
}
\value{
The MGSM27 signature matrix from Identifying a High-risk Cellular Signature in the Multiple Myeloma Bone Marrow Microenvironment
}
\description{
MGSM27 <- loadMGSM27()
}
