% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeSigMatrix.R
\name{loadModMap}
\alias{loadModMap}
\title{Load the LM22 xCell map}
\usage{
loadModMap()
}
\value{
A map between xCell cell type names and LM22 cell type names
}
\description{
Load the LM22 xCell map
}
\examples{
xcellMap <- loadModMap()
}
