% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runLoops.R
\name{meanResults}
\alias{meanResults}
\title{A meta analysis for the results from multiple iterations}
\usage{
meanResults(allResList, changePer = 1)
}
\arguments{
\item{allResList}{A list of results generated from all the iterative calls of testAllSigMatrices}

\item{changePer}{The maximum percentage of change allowed for convergence}
}
\value{
The mean and standard deviation of all the results, along with the number of iterations needed for the results to converge.
A meta analysis for the results from multiple iterations

The mean and standard deviation of all the results, along with the number of iterations needed for the results to converge.
}
\description{
Calculate the mean and the standard deviation of the results from all the iterations, and also 
test for convergence by % of change with each additional iteration.

Calculate the mean and the standard deviation of the results from all the iterations, and also 
test for convergence by % of change with each additional iteration.
}
