% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeSigMatrix.R
\name{missForest.par}
\alias{missForest.par}
\title{Use parallel missForest to impute missing values.
  This wrapper is required because missForest crashed if you have more cores than variables.}
\usage{
missForest.par(dataMat, parallelize = "variables")
}
\arguments{
\item{dataMat}{Columns are features, Rows examples. The data with NA values.  'xmis' in missForest}

\item{parallelize}{split on 'forests' or 'variables' (DEFAULT: 'variables')}
}
\value{
a matrxin including imputed values
}
\description{
newMatrix <- missForest.par(dataMat)
}
