% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeSigMatrix.R
\name{rankByT.oneCore}
\alias{rankByT.oneCore}
\title{Use a t-test to rank to features for each cell type.
  Single core version}
\usage{
rankByT.oneCore(geneExpr, qCut = 0.3)
}
\arguments{
\item{geneExpr}{The gene expression data}

\item{qCut}{(DEFAULT: 0.3)}
}
\value{
a list of cell types with data frames ranking genes
}
\description{
gList <- rankByT.oneCore(geneExpr, qCut=0.3)
}
