% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeSigMatrix.R
\name{remakeLM22p}
\alias{remakeLM22p}
\title{Use the full LM22 data matrix and add a few additional genes to cover osteoblasts, osteoclasts,
Plasma.memory, MM.  In many ways this is just a convenient wrapper for AugmentSigMatrix}
\usage{
remakeLM22p(exprData, fullLM22, smallLM22 = NULL, plotToPDF = TRUE,
  condTol = 1.01, postNorm = FALSE, autoDetectMin = FALSE)
}
\arguments{
\item{exprData}{The gene express data to use to augment LM22}

\item{fullLM22}{LM22 data with all genes.  Available in ADAPTSdata2::fullLM22}

\item{smallLM22}{The small LM22 matrix, is it includes new cell types in exprData those will not be overwritten (DEFAULT: NULL, i.e. buildLM22plus(useLM22genes = TRUE)}

\item{plotToPDF}{TRUE: pdf, FALSE: standard display (DEFAULT: TRUE)}

\item{condTol}{The tolerance in the reconstruction algorithm.  1.0 = no tolerance, 1.05 = 5\% tolerance (DEFAULT: 1.01)}

\item{postNorm}{Set to TRUE to normalize new signatures to match old signatures.  To Do: Redo Kappa curve? (DEFAULT: FALSE)}

\item{autoDetectMin}{Set to true to automatically detect the first local minima. GOOD PRELIMINARY RESULTS (DEAFULT: FALSE)}
}
\value{
a cell type signature matrix
}
\description{
matData <- remakeLM22p(exprData, fullLM22)
}
