% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/onlyDeconAlgorithms.R
\name{weightNorm}
\alias{weightNorm}
\title{SVMDECONV helper function}
\usage{
weightNorm(w)
}
\arguments{
\item{w}{The weight vector from fitting an SVM, something like something like t(fit1$coefs) \%*\% fit1$SV, where fit comes from  <- svm(m~B, nu=0.25, kernel="linear"))}
}
\value{
a weight vector
}
\description{
Use weightNorm to normalize the SVM weights.  Used for SVMDECONV

w1 <- weightNorm(w)
}
